added *4 to all unspecified integer declarations
git-svn-id: svn://svn.code.sf.net/p/irbem/code/trunk@434 80555d2d-b7e3-4ae1-aded-73e86dfc2014

+++ source/AFRL_CRRES_models.f
@@ -787,7 +787,7 @@ c
 c
        IMPLICIT NONE
 c
-       INTEGER ish
+       INTEGER*4 ish
        INTEGER*4 Ne,Nbb0,Nl
 c
        REAL*8 Ec(22),tmp_Ec(22)
@@ -841,7 +841,7 @@ c
 c
        IMPLICIT NONE
 c
-       INTEGER ish
+       INTEGER*4 ish
        INTEGER*4 Ne,Nbb0,Nl
 c
        REAL*8 Ec(22),tmp_Ec(22)
